[Feat] Diary 일기 기능 추가

+++ src/main/java/com/example/LOLDiary/domain/summoner/SummonerRepository.java
@@ -3,6 +3,6 @@ package com.example.LOLDiary.domain.summoner;
 import com.example.LOLDiary.domain.summoner.Summoner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SummonerRepository extends JpaRepository<Summoner, Integer> {
+public interface SummonerRepository extends JpaRepository<Summoner, Long> {
 
 }
